Rename all instances of `PaymentRequest` to `PaymentTerms` & refactoring to place each of the three dpp message type structs in their own folders
